# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "FullStackTest Api",
        "description": "Client/Providers CRUD",
        "version": "1.0.0"
    },
    "host": "localhost",
    "schemes": [
        "http"
    ],
    "basePath": "/",
    "produces": [
        "application/json"
    ],
    "paths": {
        "/clients": {
            "get": {
                "summary": "Client's list",
                "description": "The clients endpoint returns the client's list and their providers.\n",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients list"
                ],
                "responses": {
                    "200": {
                        "description": "An array of clients objects",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Client"
                            }
                        }
                    }
                }
            }
        },
        "/clients/:id": {
            "post": {
                "summary": "Updates an existing client or creates a new one",
                "description": "This endpoint lets you create or update clients and their providers\n",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "client",
                        "in": "body",
                        "description": "Client object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ClientParamResponse"
                        }
                    }
                ],
                "tags": [
                    "Creating a client, updating a client"
                ],
                "responses": {
                    "200": {
                        "description": "A json object with the results of the operations",
                        "schema": {
                            "type": "object",
                            "items": {
                                "$ref": "#/definitions/ClientParamResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "That email is already registered",
                        "schema": {
                            "type": "object",
                            "items": {
                                "$ref": "#/definitions/ErrorResponse"
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete an existing client",
                "description": "This endpoint lets you delete a specific client.\n",
                "parameters": [
                    {
                        "name": "params",
                        "in": "query",
                        "description": "ID of client to be deleted",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "Deleting clients"
                ],
                "responses": {
                    "200": {
                        "description": "The client has been deleted or it wasn't exists"
                    }
                }
            }
        },
        "/providers": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Provider's list",
                "description": "The providers endpoint returns the provider's list.\n",
                "tags": [
                    "Providers list"
                ],
                "responses": {
                    "200": {
                        "description": "An array of providers objects",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ProviderParamResponse"
                            }
                        }
                    }
                }
            }
        },
        "/providers/:id": {
            "post": {
                "summary": "Updates an existing provider or creates a new one",
                "description": "This endpoint lets you create or update providers\n",
                "tags": [
                    "Creating a provider, updating a provider"
                ],
                "parameters": [
                    {
                        "name": "provider",
                        "in": "body",
                        "description": "Provider Object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Provider"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A json object with the results of the operations",
                        "schema": {
                            "type": "object",
                            "items": {
                                "$ref": "#/definitions/ProviderParamResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "items": {
                                "$ref": "#/definitions/ErrorResponse"
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete an existing provider",
                "description": "This endpoint lets you delete a specific provider.\n",
                "parameters": [
                    {
                        "name": "params",
                        "in": "query",
                        "description": "Provider's ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "Deleting providers"
                ],
                "responses": {
                    "200": {
                        "description": "The provider has been deleted or it wasn't exists."
                    }
                }
            }
        }
    },
    "definitions": {
        "Client": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "description": "Client's name."
                },
                "phone": {
                    "type": "string",
                    "description": "Client's phone."
                },
                "email": {
                    "type": "string",
                    "description": "Client's email.,"
                },
                "client_providers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ClientProvider"
                    },
                    "description": "An array containing the providers's id's associated with each client"
                }
            }
        },
        "ClientParamResponse": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "boolean",
                    "description": "True if a new record was created"
                },
                "model": {
                    "$ref": "#/definitions/Client"
                }
            }
        },
        "ProviderParamResponse": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "boolean",
                    "description": "True if was a new record"
                },
                "model": {
                    "$ref": "#/definitions/Provider"
                }
            }
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "string",
                    "description": "Description of the error"
                }
            }
        },
        "Provider": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "description": "Unique identifier representing a specific provider."
                },
                "name": {
                    "type": "string",
                    "description": "The provider's name."
                }
            }
        },
        "ClientProvider": {
            "type": "object",
            "properties": {
                "delete": {
                    "type": "boolean",
                    "description": "Usefull to identify if the provider needs to be removed from client provider's list."
                },
                "id": {
                    "type": "integer",
                    "description": "Provider's unique identifier representing a specific provider. It's and autonumeric field."
                }
            }
        }
    }
}